requirements.txt
paho-mqtt>=1.6.0
PyGObject>=3.42.0